requirements.txt
sqlalchemy
pandas
psycopg2-binary
huggingface_hub
python-dotenv
streamlit
datasets
evaluate